fix(feed): correct site `url` in feed metadata
When following the existing RSS feed, none of the generated URLs
are correct, which means following a link to a post in a reader doesn't
work.

Not only does this add the needed slashes (`//`), but also points to the
HTTPS URL for the site's primary domain, rather than GitHub Pages.